Update the `HPX_GAS_BLOCK_BYTES_MAX` to be a GAS-specific value, and extend the GAS interface to support `size_t` block sizes.
The HPX_GAS_BLOCK_BYTES_MAX can be used by the application to decompose data
structures in an application specific way, but it depends on the GAS that has
been selected dynamically. For example, our PGAS implementation limits blocks to
2^32 bytes, but the SMP implementation has no obvious bound.

In order to use this effectively, the block-size interface needs to be updated
everywhere to use `size_t` block sizes as well. Specific implementations that
need smaller block sizes can test dynamically and then do a narrowing cast.

+++ libhpx/gas/pgas/pgas.h
@@ -62,7 +62,7 @@ extern HPX_ACTION_DECL(pgas_free);
 ///
 /// @returns            A global address representing the base of the
 ///                     allocation, or HPX_NULL if there is an error.
-hpx_addr_t pgas_alloc_cyclic_sync(size_t n, uint32_t bsize);
+hpx_addr_t pgas_alloc_cyclic_sync(size_t n, size_t bsize);
 
 /// Synchronous entry point for calloc.
 ///
@@ -71,7 +71,7 @@ hpx_addr_t pgas_alloc_cyclic_sync(size_t n, uint32_t bsize);
 ///
 /// @returns            A global address representing the base of the
 ///                     allocation, or HPX_NULL if there is an error.
-hpx_addr_t pgas_calloc_cyclic_sync(size_t n, uint32_t bsize);
+hpx_addr_t pgas_calloc_cyclic_sync(size_t n, size_t bsize);
 
 /// @}
